uwb: sr100: use edge type interrupt
sr100 pulls up the interrupt line twice during read io. The first
assertion happens when there's something to be read, and next irq
followed after a host sent a signal back to a uwb chip.

Currently the driver is using gpio polling for tracking the irq
line state. This could be problematic when the system is busy.

This patch changes the irq type to rising edge and tracks the irq
state in irq handler instead. And change the read() back to
use non-timedout wait_event() call, which is the same mechanism
used by NXP original implementation.

Bug: 235749999
Test: UWB factory test
Signed-off-by: Ikjoon Jang <ikjn@google.com>
Change-Id: I0416363fe26121176e3e9c027f963c9d875ffd04